refactor: add hook transform workspace
