+++ apps/web/src/plugins/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-import { jotaiStore, jotaiWorkspacesAtom } from '../atoms';
 import {
   BlockSuiteWorkspace,
   FlavourToWorkspace,
@@ -64,27 +63,3 @@ export const WorkspacePlugins = {
 } satisfies {
   [Key in RemWorkspaceFlavour]: WorkspacePlugin<Key>;
 };
-
-/**
- * Transform workspace from one flavour to another
- *
- * The logic here is to delete the old workspace and create a new one.
- */
-export async function transformWorkspace<
-  From extends RemWorkspaceFlavour,
-  To extends RemWorkspaceFlavour
->(from: From, to: To, workspace: FlavourToWorkspace[From]): Promise<string> {
-  // fixme: type cast
-  await WorkspacePlugins[from].CRUD.delete(workspace as any);
-  const newId = await WorkspacePlugins[to].CRUD.create(
-    workspace.blockSuiteWorkspace
-  );
-  const workspaces = jotaiStore.get(jotaiWorkspacesAtom);
-  const idx = workspaces.findIndex(ws => ws.id === workspace.id);
-  workspaces.splice(idx, 1, {
-    id: newId,
-    flavour: to,
-  });
-  jotaiStore.set(jotaiWorkspacesAtom, [...workspaces]);
-  return newId;
-}
